+++ components/fxa-client/src/internal/mod.rs
@@ -11,7 +11,7 @@ use self::{
     state_persistence::PersistedState,
     telemetry::FxaTelemetry,
 };
-use crate::{Error, FxaConfig, Result};
+use crate::{Error, FxaConfig, FxaRustAuthState, Result};
 use serde_derive::*;
 use std::{
     collections::{HashMap, HashSet},
@@ -77,10 +77,12 @@ impl FirefoxAccount {
             last_handled_command: None,
             commands_data: HashMap::new(),
             device_capabilities: HashSet::new(),
+            server_local_device_info: None,
             session_token: None,
             current_device_id: None,
             last_seen_profile: None,
             access_token_cache: HashMap::new(),
+            logged_out_from_auth_issues: false,
         })
     }
 
@@ -188,6 +190,10 @@ impl FirefoxAccount {
         }
     }
 
+    pub fn get_auth_state(&self) -> FxaRustAuthState {
+        self.state.get_auth_state()
+    }
+
     /// Disconnect from the account and optionally destroy our device record. This will
     /// leave the account object in a state where it can eventually reconnect to the same user.
     /// This is a "best effort" infallible method: e.g. if the network is unreachable,
@@ -222,6 +228,26 @@ impl FirefoxAccount {
         self.state.disconnect();
         self.clear_devices_and_attached_clients_cache();
         self.telemetry = FxaTelemetry::new();
+    }
+
+    /// Update the state based on authentication issues.
+    ///
+    /// **💾 This method alters the persisted account state.**
+    ///
+    /// Call this if you know there's an authentication / authorization issue that requires the
+    /// user to re-authenticated.  It transitions the user to the [FxaRustAuthState.AuthIssues] state.
+    pub fn on_auth_issues(&mut self) {
+        self.state.on_auth_issues();
+        self.clear_devices_and_attached_clients_cache();
+        self.telemetry = FxaTelemetry::new();
+    }
+
+    pub fn simulate_temporary_auth_token_issue(&mut self) {
+        self.state.simulate_temporary_auth_token_issue()
+    }
+
+    pub fn simulate_permanent_auth_token_issue(&mut self) {
+        self.state.simulate_permanent_auth_token_issue()
     }
 }
 
@@ -496,6 +522,66 @@ mod tests {
         assert!(fxa.state.refresh_token().is_some());
         fxa.disconnect();
         assert!(fxa.state.refresh_token().is_none());
+    }
+
+    #[test]
+    fn test_on_auth_issues() {
+        let config = Config::new("https://stable.dev.lcip.org", "12345678", "https://foo.bar");
+        let mut fxa = FirefoxAccount::with_config(config);
+
+        fxa.state.force_refresh_token(RefreshToken {
+            token: "refresh_token".to_owned(),
+            scopes: HashSet::new(),
+        });
+        fxa.state.force_current_device_id("original-device-id");
+        assert_eq!(fxa.get_auth_state(), FxaRustAuthState::Connected);
+
+        fxa.on_auth_issues();
+        assert_eq!(fxa.get_auth_state(), FxaRustAuthState::AuthIssues);
+
+        fxa.state.complete_oauth_flow(
+            vec![],
+            RefreshToken {
+                token: "refreshtok".to_owned(),
+                scopes: HashSet::default(),
+            },
+            None,
+        );
+        assert_eq!(fxa.get_auth_state(), FxaRustAuthState::Connected);
+
+        // The device ID should be the same as before `on_auth_issues` was called.  This
+        // way, methods like `ensure_capabilities` and `set_device_name`, can re-use it and we
+        // won't try to create a new device record.
+        assert_eq!(fxa.state.current_device_id(), Some("original-device-id"));
+    }
+
+    #[test]
+    fn test_get_auth_state() {
+        let config = Config::new("https://stable.dev.lcip.org", "12345678", "https://foo.bar");
+        let mut fxa = FirefoxAccount::with_config(config);
+
+        fn assert_auth_state(fxa: &FirefoxAccount, correct_state: FxaRustAuthState) {
+            assert_eq!(fxa.get_auth_state(), correct_state);
+
+            let persisted = FirefoxAccount::from_json(&fxa.to_json().unwrap()).unwrap();
+            assert_eq!(persisted.get_auth_state(), correct_state);
+        }
+
+        // The state starts as disconnected
+        assert_auth_state(&fxa, FxaRustAuthState::Disconnected);
+
+        // When we get the refresh tokens the state changes to connected
+        fxa.state.force_refresh_token(RefreshToken {
+            token: "refresh_token".to_owned(),
+            scopes: HashSet::new(),
+        });
+        assert_auth_state(&fxa, FxaRustAuthState::Connected);
+
+        fxa.disconnect();
+        assert_auth_state(&fxa, FxaRustAuthState::Disconnected);
+
+        fxa.disconnect();
+        assert_auth_state(&fxa, FxaRustAuthState::Disconnected);
     }
 
     #[test]